+++ examples/hello/lib/app.dart
@@ -1,9 +1,0 @@
-// Copyright (c) 2020, the Dart project authors.
-// Please see the AUTHORS file or details. Use of this source code is
-// governed by a BSD-style license that can be found in the LICENSE file.
-
-import 'package:functions_framework/functions_framework.dart';
-import 'package:shelf/shelf.dart';
-
-@CloudFunction()
-Response handleGet(Request request) => Response.ok('Hello, World!');